backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-decouple>=3.8,<4.0
requests>=2.31,<3.0
django-cors-headers>=4.3,<5.0

# Production
gunicorn>=21.0,<23.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.6,<7.0
dj-database-url>=2.1,<3.0
